refactor: reorganize HTTP-related classes and implement CORS middleware
