+++ src/Http/Middleware/MiddlewareInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Fluxoft\Rebar\Http\Middleware;
+
+interface MiddlewareInterface {
+	public function Process(Request $request, Response $response, callable $next): Response;
+}
